v1.1 lab3

+++ Dyryaev_Vsevolod_381903-2/3/include/full_analiz.h
@@ -6,5 +6,5 @@
 class Full_a
 {
 public:
-    int full_analiz(std::string);
+    float full_analiz(std::string);
 };
